Use the main method and Introduce some level of encapsulation
Even a file meant to be used as a script should be importable and a mere import should not have the side effect of executing the script's main functionality. The main functionality should be in a ```main()``` function.
In Python, pydoc as well as unit tests require modules to be importable. Your code should always check ```if __name__ == '__main__'``` before executing your main program so that the main program is not executed when the module is imported.
```
def main():
      ...

if __name__ == '__main__':
    main()
```
https://google.github.io/styleguide/pyguide.html
https://www.python.org/dev/peps/pep-0299/