Read ability data from sqlite3
There's a large room for improvement, but this is good for a base 🙂
All the data is stored in revoDB.db, the exact same data as the previous lists. For now, on startup, the lists are populated from the database (For the user's abilities only!)

This change also made the ability names use pascal case with spaces, which made the output cleaner
e.g. instead of [CLEAVE, SLAUGHTER, ASSAULT]
this: [Cleave, Slaughter, Assault]